Add travis integration, fix build
This resolves #3 by adding a Travis configuration. Additionally, I
realized that the build doesn't work fresh, so I added some resolutions
to the bower file (does anybody know of a better thing to do? I always
feel dirty doing that).

Additionally, I corrected added a license field to `package.json` and
corrected the format of the one in `bower.json`.